Validator fixes.
