+++ src/Validators/Test/MaxLengthValidatorTest.php
@@ -1,22 +1,23 @@
 <?php
 
-namespace PDGA\DataObjects\Validators;
+namespace PDGA\DataObjects\Validators\Test;
 
+use PDGA\DataObjects\Validators\MaxLengthValidator;
 use PHPUnit\Framework\TestCase;
 
-class MinLengthValidatorTest extends TestCase
+class MaxLengthValidatorTest extends TestCase
 {
-    private MinLengthValidator $validator;
+    private MaxLengthValidator $validator;
 
     /**
-     * All tests are against a min length of 5.
+     * All tests are against a max length of 5.
      * @var int
      */
-    private int $min_length = 5;
+    private int $max_length = 5;
 
     public function setUp(): void
     {
-        $this->validator = new MinLengthValidator($this->min_length);
+        $this->validator = new MaxLengthValidator($this->max_length);
     }
 
     public function testGetErrorMessage()
@@ -24,7 +25,7 @@ class MinLengthValidatorTest extends TestCase
         $propName = 'Property';
 
         $this->assertEquals(
-            "Minimum length of $propName is $this->min_length characters.",
+            "Maximum length of $propName is $this->max_length characters.",
             $this->validator->getErrorMessage($propName)
         );
     }
@@ -38,16 +39,16 @@ class MinLengthValidatorTest extends TestCase
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testStringExceedsMinLength()
+    public function testStringExceedsMaxLength()
     {
         // Use a string value longer than 5 characters.
         $value = "longerthanfive";
 
-        // The value should validate.
-        $this->assertEquals(true, $this->validator->validate($value));
+        // The value should not validate.
+        $this->assertEquals(false, $this->validator->validate($value));
     }
 
-    public function testStringEqualsMinLength()
+    public function testStringEqualsMaxLength()
     {
         // Use a string of 5 characters.
         $value = "bang!";
@@ -56,25 +57,25 @@ class MinLengthValidatorTest extends TestCase
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testStringUnderMinLength()
+    public function testStringUnderMaxLength()
     {
         // Use a string of 4 characters.
         $value = "four";
-
-        // The value should not validate.
-        $this->assertEquals(false, $this->validator->validate($value));
-    }
-
-    public function testIntegerExceedsMinLength()
-    {
-        // Use an integer longer than 5 characters.
-        $value = 123456;
 
         // The value should validate.
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testIntegerEqualsMinLength()
+    public function testIntegerExceedsMaxLength()
+    {
+        // Use an integer longer than 5 characters.
+        $value = 123456;
+
+        // The value should not validate.
+        $this->assertEquals(false, $this->validator->validate($value));
+    }
+
+    public function testIntegerEqualsMaxLength()
     {
         // Use an integer of 5 characters.
         $value = 12345;
@@ -83,25 +84,25 @@ class MinLengthValidatorTest extends TestCase
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testIntegerUnderMinLength()
+    public function testIntegerUnderMaxLength()
     {
         // Use an integer shorter than 5 characters.
         $value = 1234;
-
-        // The value should not validate.
-        $this->assertEquals(false, $this->validator->validate($value));
-    }
-
-    public function testArrayExceedsMinLength()
-    {
-        // Use an array of more than 5 items.
-        $value = [1, 2, 3, 4, 5, 6];
 
         // The value should validate.
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testArrayEqualsMinLength()
+    public function testArrayExceedsMaxLength()
+    {
+        // Use an array of more than 5 items.
+        $value = [1, 2, 3, 4, 5, 6];
+
+        // The value should not validate.
+        $this->assertEquals(false, $this->validator->validate($value));
+    }
+
+    public function testArrayEqualsMaxLength()
     {
         // Use an array of 5 items.
         $value = [1, 2, 3, 4, 5];
@@ -110,22 +111,22 @@ class MinLengthValidatorTest extends TestCase
         $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testArrayUnderMinLength()
+    public function testArrayUnderMaxLength()
     {
         // Use an array of fewer than 5 items.
         $value = [1, 2, 3, 4];
 
-        // The value should not validate.
-        $this->assertEquals(false, $this->validator->validate($value));
+        // The value should validate.
+        $this->assertEquals(true, $this->validator->validate($value));
     }
 
-    public function testMultiDimensionalArrayUnderMinLength()
+    public function testMultiDimensionalArrayUnderMaxLength()
     {
         // Use an array of fewer than 5 items at the first dimension
         // with more than 5 items total in all dimensions.
         $value = [1, 2, 3, 4 => [5, 6, 7, 8]];
 
-        // The value should not validate.
-        $this->assertEquals(false, $this->validator->validate($value));
+        // The value should validate.
+        $this->assertEquals(true, $this->validator->validate($value));
     }
 }
