+++ src/Validators/Test/ZipValidatorTest.php
@@ -1,8 +1,9 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+namespace PDGA\DataObjects\Validators\Test;
 
 use PDGA\DataObjects\Validators\ZipValidator;
+use PHPUnit\Framework\TestCase;
 
 class ZipValidatorTest extends TestCase
 {
